fix seguridad al eliminar

+++ eliminar.php
@@ -1,5 +1,12 @@
 <?php
-include_once  "conexion.php";
+session_start();
+
+if (!isset($_SESSION['usuario'])) {
+    header("location: index.php");
+    exit();
+}
+
+include_once "conexion.php";
 $conn = getConexion('pokemon');
 $sql = "DELETE FROM pokemones WHERE numero = ?";
